small change to getNearRadius (changed the cast to float) removed unused imports for copy particle

+++ src/simulation/body/particle/AttractorParticle.java
@@ -25,7 +25,7 @@ public class AttractorParticle extends Particle implements Gravity {
   @Override
   public short getNearRadius() {
     return (short) (
-      Settings.get(Settings.Constants.MAX_RADIUS) + (float) strength / 5
+      Settings.get(Settings.Constants.MAX_RADIUS) + strength / 5f
     );
   }
